# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * PackSmart AI decision-support API
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardSummaryTopMaterialsItem } from './dashboardSummaryTopMaterialsItem';
import type { RecommendationSummary } from './recommendationSummary';

export interface DashboardSummary {
  totalRecommendations: number;
  recentRecommendations: number;
  savedProducts: number;
  averageShelfLife: string;
  topMaterials: DashboardSummaryTopMaterialsItem[];
  latest: RecommendationSummary[];
}