# lib/api-zod/src/generated/types/publicShare.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * PackSmart AI decision-support API
 * OpenAPI spec version: 0.1.0
 */

export interface PublicShare {
  shareId: string;
  url: string;
}