# lib/api-zod/src/generated/types/recommendationInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * PackSmart AI decision-support API
 * OpenAPI spec version: 0.1.0
 */
import type { RecommendationInputPreferences } from './recommendationInputPreferences';
import type { RecommendationInputProperties } from './recommendationInputProperties';
import type { RecommendationInputStorage } from './recommendationInputStorage';
import type { RecommendationInputTransportation } from './recommendationInputTransportation';

export interface RecommendationInput {
  /** @minLength 1 */
  commodityName: string;
  /** @minLength 1 */
  category: string;
  properties: RecommendationInputProperties;
  storage: RecommendationInputStorage;
  transportation: RecommendationInputTransportation;
  preferences: RecommendationInputPreferences;
}