# lib/api-zod/src/generated/types/recommendationSaveInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * PackSmart AI decision-support API
 * OpenAPI spec version: 0.1.0
 */
import type { RecommendationInput } from './recommendationInput';
import type { RecommendationResult } from './recommendationResult';

export interface RecommendationSaveInput {
  input: RecommendationInput;
  result: RecommendationResult;
}